third_party/blink/web_tests/wpt_internal/css/css-anchor-position/anchor-size-minmax-001.html

<!DOCTYPE html>
<link rel="help" href="https://tabatkins.github.io/specs/css-anchor-position/#propdef-anchor-name">
<link rel="author" href="mailto:kojii@chromium.org">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/check-layout-th.js"></script>
<style>
.container {
  position: relative;
  height: 10px;
}
.anchor1 {
  anchor-name: --a1;
  width: 5px;
  height: 7px;
  background: orange;
}
.target {
  position: absolute;
}
</style>
<body onload="checkLayout('.target')">
  <div class="container">
    <div class="anchor1"></div>

    <div class="target"
      style="min-width: anchor-size(--a1 width)"
      data-expected-width=5></div>

    <div class="target"
      style="max-width: anchor-size(--a1 width)"
      data-expected-width=5>
      <div style="width: 100px"></div>
    </div>
  </div>
</body>
